@@ -1,8 +1,4 @@
-import Vue from 'vue';
-import App from './App.vue';
+import { createApp } from "vue";
+import App from "./App.vue";
 
-// Create our Vue App and replace the <div id="app"></div> with it
-new Vue({
-  el: '#app',
-  render: h => h(App)
-});
+createApp(App).mount("#app");
